requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cloudinary==1.29.0
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
Pillow==9.1.1
pkg_resources==0.0.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
